Moved aspnet host into tests project
