+++ sdk/extensions/Azure.Extensions.sln
@@ -39,8 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Azure.WebJobs.Ext
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Extensions.Azure.Samples", "Microsoft.Extensions.Azure\samples\Microsoft.Extensions.Azure.Samples.csproj", "{D131D37E-9E4E-4511-BBB1-28F9DC8C76CC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.Extensions.Azure.Tests.AspNetHost", "Microsoft.Extensions.Azure\tests\aspnet-host\Microsoft.Extensions.Azure.Tests.AspNetHost.csproj", "{78C229CF-B8E8-43D4-AB81-41273E56E6D0}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -231,18 +229,6 @@ Global
 		{D131D37E-9E4E-4511-BBB1-28F9DC8C76CC}.Release|x64.Build.0 = Release|Any CPU
 		{D131D37E-9E4E-4511-BBB1-28F9DC8C76CC}.Release|x86.ActiveCfg = Release|Any CPU
 		{D131D37E-9E4E-4511-BBB1-28F9DC8C76CC}.Release|x86.Build.0 = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|x64.Build.0 = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Debug|x86.Build.0 = Debug|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|Any CPU.Build.0 = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|x64.ActiveCfg = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|x64.Build.0 = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|x86.ActiveCfg = Release|Any CPU
-		{78C229CF-B8E8-43D4-AB81-41273E56E6D0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
