+++ sdk/extensions/Microsoft.Extensions.Azure/tests/Microsoft.Extensions.Azure.Tests.csproj
@@ -3,11 +3,6 @@
     <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
   </PropertyGroup>
 
-  <!-- Do not include the ASPNET Core test host project -->
-  <ItemGroup>
-    <Compile Remove="aspnet-host\**" />
-  </ItemGroup>
-
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.DependencyInjection" VersionOverride="8.0.1" />
     <PackageReference Include="Microsoft.Extensions.PlatformAbstractions" />
@@ -23,14 +18,29 @@
     <ProjectReference Include="..\src\Microsoft.Extensions.Azure.csproj" />
   </ItemGroup>
 
-  <!-- Include samples for testing compatibility with the ASPNET integration framework on net8+ -->
+  <!-- Include ASPNET content for testing compatibility with the integration framework on net8+ only -->
   <ItemGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net8.0'))">
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+    <PackageReference Include="Azure.Security.KeyVault.Secrets" />
 
     <!--
       Override the ASPNET Core testing package locally until the 8.x bump in the engineering system.
       Tracked by: #48425
     -->
-    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing"  VersionOverride="8.0.13"/>
-    <ProjectReference Include=".\aspnet-host\Microsoft.Extensions.Azure.Tests.AspNetHost.csproj" />
+    <PackageReference Include="Microsoft.AspNetCore.Mvc.Testing" VersionOverride="8.0.13" />
+  </ItemGroup>
+
+  <!-- Don't include the generated entry point for net8+ because one is explicitly defined in the ASPNET content -->
+  <PropertyGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net8.0'))">
+    <GenerateProgramFile>false</GenerateProgramFile>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <None Update="aspnet-host\appsettings.Development.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="aspnet-host\appsettings.json">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
   </ItemGroup>
 </Project>
